chore: misc React template improvements
- Bump dependencies
- Simplify config
- Restructure app so that loading indicator doesn't obscure chain switcher
