+++ templates/react-papi-tailwind/src/App.tsx
@@ -1,36 +1,43 @@
 import { type ChainId } from "@reactive-dot/core";
-
+import { ChainProvider, ReactiveDotProvider } from "@reactive-dot/react";
 import { ConnectionButton } from "dot-connect/react.js";
+import { Suspense, useState } from "react";
 
 import "./App.css";
-import { ChainSwitch } from "./components/ChainSwitch";
 import { ChainPage } from "./ChainPage";
-
 import polkadotLogo from "./assets/polkadot-logo.svg";
-
 import { AccountList } from "./components/AccountList";
+import { ChainSwitch } from "./components/ChainSwitch";
+import Loading from "./components/Loading";
+import { config } from "./reactive-dot";
 
-function App({ chainId, setChainId }: { chainId: ChainId; setChainId: React.Dispatch<React.SetStateAction<ChainId>> }) {
+function App() {
+  const [chainId, setChainId] = useState<ChainId>("paseo");
+
   return (
-    <>
-      <div className="fixed right-10 top-10">
-        <div className="ml-5 inline-block">
-          <ConnectionButton />
+    <ReactiveDotProvider config={config}>
+      <ChainProvider chainId={chainId}>
+        <div className="fixed right-10 top-10">
+          <div className="ml-5 inline-block">
+            <ConnectionButton />
+          </div>
+          <div className="ml-5 inline-block">
+            <ChainSwitch chainId={chainId} setChainId={setChainId} />
+          </div>
         </div>
-        <div className="ml-5 inline-block">
-          <ChainSwitch chainId={chainId} setChainId={setChainId} />
-        </div>
-      </div>
 
-      <img src={polkadotLogo} className="logo mx-auto h-52 p-4" alt="Polkadot logo" />
-      <div className="container mx-auto p-2 leading-6">
-        <ChainPage />
-      </div>
+        <Suspense fallback={<Loading />}>
+          <img src={polkadotLogo} className="logo mx-auto h-52 p-4" alt="Polkadot logo" />
+          <div className="container mx-auto p-2 leading-6">
+            <ChainPage />
+          </div>
 
-      <div>
-        <AccountList />
-      </div>
-    </>
+          <div>
+            <AccountList />
+          </div>
+        </Suspense>
+      </ChainProvider>
+    </ReactiveDotProvider>
   );
 }
