--- templates/react-papi-tailwind/src/reactive-dot-types.d.ts
@@ -1,7 +1,0 @@
-import type { config } from "./reactive-dot.ts";
-
-declare module "@reactive-dot/core" {
-  export interface Register {
-    config: typeof config;
-  }
-}
