+++ templates/react-papi-tailwind/src/reactive-dot.ts
@@ -8,33 +8,6 @@ import { registerDotConnect } from "dot-connect";
 
 const lightClientProvider = createLightClientProvider();
 
-const wallets = [
-  new InjectedWalletProvider(),
-  new LedgerWallet(),
-  // Uncomment to configure WalletConnect.
-  //new WalletConnect({
-  //  projectId: "WALLET_CONNECT_PROJECT_ID",
-  //  providerOptions: {
-  //    metadata: {
-  //      name: "APP_NAME",
-  //      description: "APP_DESCRIPTION",
-  //      url: "APP_URL",
-  //      icons: ["APP_ICON"]
-  //    }
-  //  },
-  //  // https://github.com/ChainAgnostic/CAIPs/blob/main/CAIPs/caip-13.md
-  //  chainIds: [
-  //    "polkadot:91b171bb158e2d3848fa23a9f1c25182" // Polkadot
-  //  ],
-  //  optionalChainIds: [
-  //    "polkadot:91b171bb158e2d3848fa23a9f1c25182", // Polkadot
-  //    "polkadot:b0a8d493285c2df73290dfb7e61f870f", // Kusama
-  //    "polkadot:77afd6190f1554ad45fd0d31aee62aac", // Paseo
-  //    "polkadot:e143f23803ac50e8f6f8e62695d1ce9e" // Westend
-  //  ]
-  //})
-];
-
 const polkadotChain = lightClientProvider.addRelayChain({ id: "polkadot" });
 const polkadotAHChain = polkadotChain.addParachain({ id: "polkadot_asset_hub" });
 
@@ -60,7 +33,38 @@ export const config = defineConfig({
       provider: paseoAHChain,
     },
   },
-  wallets,
+  wallets: [
+    new InjectedWalletProvider(),
+    new LedgerWallet(),
+    // Uncomment to configure WalletConnect.
+    //new WalletConnect({
+    //  projectId: "WALLET_CONNECT_PROJECT_ID",
+    //  providerOptions: {
+    //    metadata: {
+    //      name: "APP_NAME",
+    //      description: "APP_DESCRIPTION",
+    //      url: "APP_URL",
+    //      icons: ["APP_ICON"]
+    //    }
+    //  },
+    //  // https://github.com/ChainAgnostic/CAIPs/blob/main/CAIPs/caip-13.md
+    //  chainIds: [
+    //    "polkadot:91b171bb158e2d3848fa23a9f1c25182" // Polkadot
+    //  ],
+    //  optionalChainIds: [
+    //    "polkadot:91b171bb158e2d3848fa23a9f1c25182", // Polkadot
+    //    "polkadot:b0a8d493285c2df73290dfb7e61f870f", // Kusama
+    //    "polkadot:77afd6190f1554ad45fd0d31aee62aac", // Paseo
+    //    "polkadot:e143f23803ac50e8f6f8e62695d1ce9e" // Westend
+    //  ]
+    //})
+  ],
 });
 
-registerDotConnect({ wallets });
+declare module "@reactive-dot/core" {
+  export interface Register {
+    config: typeof config;
+  }
+}
+
+registerDotConnect(config);
